# Support wildcards with transformation targets