--- a/.ps/after-cd.ps1
+++ b/.ps/after-cd.ps1
@@ -1,4 +1,0 @@
-Write-Host
-Write-Host "Recent Branches:"
-Write-Host
-git branch --list --sort=-committerdate --format='%(committerdate:short) %(refname:short)' | select -first 20 | Write-Host