--- a/v2/tools/generator/internal/config/always_matcher.go
+++ b/v2/tools/generator/internal/config/always_matcher.go
@@ -29,3 +29,8 @@ func (a alwaysMatcher) Matches(value string) bool {
 func (a alwaysMatcher) WasMatched() error {
 	return nil
 }
+
+// IsRestrictive returns false because the always matcher doesn't restrict anything
+func (a alwaysMatcher) IsRestrictive() bool {
+	return false
+}