--- a/v2/tools/generator/internal/config/field_matcher_test.go
+++ b/v2/tools/generator/internal/config/field_matcher_test.go
@@ -44,3 +44,32 @@ func TestFieldMatcher_DeserializedFromYaml_GivesExpectedMatchResult(t *testing.T
 type matcherHost struct {
 	Field FieldMatcher `yaml:"field"`
 }
+
+func TestFieldMatcher_IsRestrictive_GivesExpectedResults(t *testing.T) {
+	t.Parallel()
+
+	cases := []struct {
+		name        string
+		yaml     string
+		restrictive bool
+	}{
+		{"Empty literal is unrestrictive", "field: ''", false},
+		{"Wildcard is unrestrictive", "field: '*'", false},
+		{"Literal is restrictive", "field: Foo", true},
+		{"Complex wildcard is restrictive", "field: Foo*", true},
+	}
+
+	for _, c := range cases {
+		c := c
+		t.Run(
+			c.name,
+			func(t *testing.T) {
+				t.Parallel()
+				g := NewGomegaWithT(t)
+
+				var h matcherHost
+				g.Expect(yaml.Unmarshal([]byte(c.yaml), &h)).To(Succeed())
+				g.Expect(h.Field.IsRestrictive()).To(Equal(c.restrictive))
+			})
+	}
+}