… v2/tools/generator/internal/config/glob_matcher_test.go
… v2/tools/generator/internal/config/literal_matcher_test.go
… v2/tools/generator/internal/config/multi_matcher_test.go
… v2/tools/generator/internal/config/type_transformer_test.go