**What this PR does / why we need it**:

Despite the appearance of fragments like this in our configuration

``` yaml
  - name: "*"
    property: Copy
    ifType:
      group: definitions
      version: "*"
      name: ResourceCopy
      optional: true
    remove: true
    because: ... elided... 
```

We did not support wildcards within the `ifType` section; the matching on version `*` was accidentally working, due to the way our package references filtered out non-alphanumeric characters during creation.

In PR #2294 I'm removing this filtering to improve our diagnostics, breaking this match.

This PR fixes the matching to properly work with wildcards. We could do some further work to improve diagnostics, etc, but I figured this was change enough.

I've included caching to improve the performance of the transforms, and to reduce the amount of heap debris, hopefully lowering the amount of time spent doing garbage collection. Crudely measures, this seems to help:

Without caching:
```
8/64: Modify property types using configured transforms, completed in 2.162s
8/64: Modify property types using configured transforms, completed in 2.328s
```

With caching
```
8/64: Modify property types using configured transforms, completed in 1.221s
8/64: Modify property types using configured transforms, completed in 1.099s
```

**Special notes for your reviewer**:

I've left in tests I wrote during troubleshooting.

**How does this PR make you feel**:
![gif](https://media.giphy.com/media/tEQ8aAOh60us0/giphy.gif)

**If applicable**:
- [x] this PR contains tests